ZocDoc/requirements-dev.txt
# Development Requirements
# Install with: pip install -r requirements-dev.txt

# Include production requirements
-r requirements.txt

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-xdist>=3.3.0

# Code quality
black>=23.0.0
flake8>=6.1.0
isort>=5.12.0
mypy>=1.5.0
pylint>=2.17.0

# Type stubs
types-beautifulsoup4
pandas-stubs

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0
